requirements.txt
pypdf~=3.17.0
spacy~=3.5.3
fastapi~=0.96.1
requests~=2.28.1
textacy~=0.12.0
transformers~=4.32.1
python-dotenv~=0.21.0